db_agama_app/requirements.txt
PySide6
mysql-connector-python
bcrypt
